code moved to XTcpSocket.cpp
